Added a query xml shortening facility.
InterMine suffers from xml-itis. This badly affects GET requests with
long queries in them. This is another method for dealing with this issue
by allowing access to a minimal query storage service on the server,
which means we can refer to queries by id. At present these queries do
not live beyond the life-time of a session, as they live in memory. They
should be treated as very short-lived entities.


Former-commit-id: 45538100821e33f6024737165925c1e6905a4e4f [formerly 62bbe93f97da5bee5dbfa7798f135cbadb3191d8]
Former-commit-id: 3662db4cd849a29f2011b8ea995bd4dc2d0c667c